requirements.txt
numpy==1.19.5
torchvision==0.10.0
streamlit==0.88.0
torch==1.9.0
Pillow==8.3.2